# `require 'language-server_protocol'` on-demand