--- a/lib/rubocop/cli/command/lsp.rb
+++ b/lib/rubocop/cli/command/lsp.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative '../../lsp/server'
-
 module RuboCop
   class CLI
     module Command
@@ -11,6 +9,8 @@ module RuboCop
         self.command_name = :lsp
 
         def run
+          # Load on demand, `languge-server-protocol` is heavy to require.
+          require_relative '../../lsp/server'
           RuboCop::LSP::Server.new(@config_store).start
         end
       end