… lib/rubocop.rb