It's a heavy gem to require for something that may not even be used:

Old:
```
$ hyperfine -w 5 -r 20 "bundle exec rubocop Rakefile"
Benchmark 1: bundle exec rubocop Rakefile
  Time (mean ± σ):      3.061 s ±  0.087 s    [User: 2.494 s, System: 0.321 s]
  Range (min … max):    2.947 s …  3.275 s    20 runs
```

New:
```
$ hyperfine -w 5 -r 20 "bundle exec rubocop Rakefile"
Benchmark 1: bundle exec rubocop Rakefile
  Time (mean ± σ):      2.801 s ±  0.064 s    [User: 2.317 s, System: 0.235 s]
  Range (min … max):    2.710 s …  2.958 s    20 runs
```

~9% faster startup time for me

Of course, most time is still spent requiring all the various cops of rubocop itself (there are quite a few)

-----------------

Before submitting the PR make sure the following are checked:

* [x] The PR relates to *only* one subject with a clear title and description in grammatically correct, complete sentences.
* [x] Wrote [good commit messages][1].
* [ ] Commit message starts with `[Fix #issue-number]` (if the related issue exists).
* [x] Feature branch is up-to-date with `master` (if not - rebase it).
* [x] Squashed related commits together.
* [ ] Added tests.
* [x] Ran `bundle exec rake default`. It executes all tests and runs RuboCop on its own code.
* [ ] Added an entry (file) to the [changelog folder](https://github.com/rubocop/rubocop/blob/master/changelog/) named `{change_type}_{change_description}.md` if the new code introduces user-observable changes. See [changelog entry format](https://github.com/rubocop/rubocop/blob/master/CONTRIBUTING.md#changelog-entry-format) for details.

[1]: https://chris.beams.io/posts/git-commit/